Create README.md
This script prompts the user for a file to decipher, generates a table of letter frequencies and then attempts to solve it by replacing most common encrypted characters with the most frequently used characters in English text.

To use this script, place your encrypted file in the same directory as the CypherText.py file and then run the CypherText.py file with Python 3. It will ask you to enter the name of your encrypted file and once you enter that, it will print a frequency list to console and write a new file with each letter substituted.